Updates Next.js to version 16 and related dependencies
Updates Next.js to version 16, along with associated dependencies like next-auth, eslint-config-next, and prisma.

This upgrade involves migrating from the deprecated FlatCompat to the recommended defineConfig and globalIgnores for ESLint configuration and updates the ESLint configuration to align with the new standards.

It also ensures that the application benefits from the latest features, performance improvements, and security patches available in the updated libraries.

Removes unused @rushstack/eslint-patch dependency.