Implements players pagination and search
Adds pagination to the players list in the admin panel.

Implements a search functionality to filter players by name, email and phone.

Creates a skeleton component for the players table to improve loading experience.

Refactors the players list page into a table component.

Fixes to players pagination.
Fixes to Skeletons.