Adds standings table with relevant statistics
Improves the standings table by including draws ('JE') and renames the 'Dif' and 'Puntos' columns to 'DIF' and 'PTS' respectively.
Also, the skeleton loader is updated to reflect the new column.
Adds a legend to explain the meaning of each column's abbreviation.
Removes conditional rendering of the table in case no tournament is selected.