Adds tournament galleries and enhances gallery forms
Enables associating galleries with tournaments in addition to teams.

Introduces a new action to fetch tournaments for gallery creation and editing.

Updates the gallery form to include tournament selection, with logic to handle either tournament or team association.

Makes `teamId` nullable in the schema.

Sets `imagePublicID` to required in the gallery image model.