Enables sorting of matches by date and week
Adds functionality to sort matches by date and week on the admin encounters page.

This change introduces sortMatchDate and sortWeek parameters to the fetchMatchesAction,
allowing users to sort matches based on either match date or week number.

A new ClearFilters component is also introduced to remove all the search params, for a better UX.