Improves match data and caching
Enhances match data by including team image URLs and extending the cache life for matches to 'days' instead of 'hours'.

This ensures that the displayed data is more visually appealing and that the cache is valid for a longer period, reducing the frequency of data fetching. Also, the status filter in `CurrentDayMatchesAction` is updated to fetch scheduled matches.