Enables score modification for completed matches
Allows users to update the score of a match even after it has been marked as completed. This functionality reverts the old standings, applies the new scores, and updates standings accordingly. Adds a button to reveal the scores, that are hidden by default, to be modified, and wraps the update behind an alert dialog to prevent accidental score updates.